+++ src/main/java/music/analysis/feature/processor/rules/MusicalRule.java
@@ -2,6 +2,6 @@ package music.analysis.feature.processor.rules;
 
 import music.analysis.feature.processor.MelodicFeatureCounter;
 
-public interface MusicalRule extends MelodicFeatureCounter<Integer> {
+public interface MusicalRule extends MelodicFeatureCounter<Double> {
 
 }
